ux-redesign: Extract Breadcrumbs from PageRouter.js
Move the `Breadcrumb` component out of `src/components/PageRouter.js`
to `src/components/Breadcrumbs/index.js`.  This will insulate the
PageRouter.js file from changes to Breadcrumbs and vice versa.
